dbFiller ahora sube el csv completo a la base de datos

+++ dbFiller/src/main/scala/sfpsfiuba/Commons.scala
@@ -3,8 +3,8 @@ package sfpsfiuba.commons
 final case class Row(
     id: Int,
     date: String,
-    open: Int,
-    high: Int,
+    open: Float,
+    high: Float,
     cierre: Float,
     rest: String
 )
@@ -12,7 +12,7 @@ final case class Row(
 object Row {
     def apply(csvRow: String): Row = csvRow match {
         case s"$id,$date,$open,$high,$cierre,$rest" 
-            => Row(id.toInt, date, open.toInt, high.toInt, cierre.toFloat, rest)
+            => Row(id.toInt, date, open.toFloat, high.toFloat, cierre.toFloat, rest)
         case _ => null
     }
 }
